pyproject.toml
[tool.poetry]
name = "python-401-api-quickstart"
version = "0.1.0"
description = ""
authors = ["JB-Tellez <jb.tellez@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
whitenoise = "^5.3.0"
django-cors-headers = "^3.11.0"
gunicorn = "^20.1.0"
django-environ = "^0.8.1"
fire = "^0.4.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.12b0"
fire = "^0.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"